Refactor SkyConnect plugins
