+++ src/PluginManager/include/PluginManager/Connect/Connect.h
