+++ src/Plugins/Module/Formation/src/FormationWidget.cpp
@@ -60,8 +60,8 @@
 #include <Persistence/Service/FlightService.h>
 #include <Persistence/Service/AircraftService.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
-#include <PluginManager/Connect.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
+#include <PluginManager/Connect/Connect.h>
 #include <Widget/Platform.h>
 #include <Widget/UnitWidgetItem.h>
 #include "Formation.h"
