+++ src/Plugins/Module/Logbook/src/LogbookPlugin.cpp
@@ -31,7 +31,7 @@
 #include <Persistence/Service/FlightService.h>
 #include <Persistence/Service/AircraftService.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
 #include <PluginManager/Module/ModuleBaseSettings.h>
 #include "LogbookWidget.h"
 #include "LogbookSettings.h"
