+++ src/Plugins/Module/Logbook/src/LogbookWidget.cpp
@@ -64,7 +64,7 @@
 #include <Persistence/Service/FlightService.h>
 #include <Persistence/PersistenceManager.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
 #include <Widget/Platform.h>
 #include <Widget/TableDateItem.h>
 #include <Widget/TableTimeItem.h>
