+++ src/UserInterface/include/UserInterface/MainWindow.h
@@ -45,7 +45,7 @@ QT_BEGIN_NAMESPACE
 namespace Ui { class MainWindow; }
 QT_END_NAMESPACE
 
-#include <PluginManager/SkyConnectIntf.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
 
 class AboutDialog;
 class SettingsDialog;
