+++ src/UserInterface/src/MainWindow.cpp
@@ -88,8 +88,8 @@
 #include <Widget/Platform.h>
 #include <Widget/RecentFileMenu.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
-#include <PluginManager/Connect.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
+#include <PluginManager/Connect/Connect.h>
 #include <PluginManager/PluginManager.h>
 #include <PluginManager/Module/ModuleIntf.h>
 #include <PluginManager/ModuleManager.h>
