+++ src/UserInterface/src/Widget/AircraftHandleWidget.cpp
@@ -39,8 +39,8 @@
 #include <Model/AircraftHandleData.h>
 #include <Model/TimeVariableData.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
-#include <PluginManager/Connect.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
+#include <PluginManager/Connect/Connect.h>
 #include "AircraftHandleWidget.h"
 #include "ui_AircraftHandleWidget.h"
