+++ src/UserInterface/src/Widget/EngineWidget.cpp
@@ -39,8 +39,8 @@
 #include <Model/EngineData.h>
 #include <Model/TimeVariableData.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
-#include <PluginManager/Connect.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
+#include <PluginManager/Connect/Connect.h>
 #include "EngineWidget.h"
 #include "ui_EngineWidget.h"
