+++ src/UserInterface/src/Widget/LightWidget.cpp
@@ -37,8 +37,8 @@
 #include <Model/LightData.h>
 #include <Model/TimeVariableData.h>
 #include <PluginManager/SkyConnectManager.h>
-#include <PluginManager/SkyConnectIntf.h>
-#include <PluginManager/Connect.h>
+#include <PluginManager/Connect/SkyConnectIntf.h>
+#include <PluginManager/Connect/Connect.h>
 #include "LightWidget.h"
 #include "ui_LightWidget.h"
